+++ src/components/orders/Orders.js
@@ -127,11 +127,9 @@ class Orders extends Component {
           }
         }
 
-        setTimeout(() => {
+        // setTimeout(() => {
           this.setState({
             data: response.data,
-            refreshing: false,
-            loading: false,
             empty: false,
             finish: true,
           });
@@ -139,13 +137,11 @@ class Orders extends Component {
           if (!noScroll) {
             this._scrollToTop(0);
           }
-        }, delay || 0);
+        // }, delay || 0);
       } else {
         setTimeout(() => {
           this.setState({
-            loading: false,
             data: null,
-            refreshing: false,
           });
         }, delay || 0);
       }
@@ -156,6 +152,10 @@ class Orders extends Component {
       store.setDeepLinkData(null);
       appConfig.device.isIOS &&
         StatusBar.setNetworkActivityIndicatorVisible(false);
+      this.setState({
+        loading: false,
+        refreshing: false,
+      });
     }
   }
 
@@ -305,11 +305,14 @@ class Orders extends Component {
   }
 
   forceUpdateOrders() {
-    this.autoUpdateDisposer = autorun(() => {
-      if (store.isUpdateOrders && !this.state.refreshing) {
-        setTimeout(() => this._getData(0, true));
-      }
-    });
+    this.autoUpdateDisposer = reaction(
+      () => store.isUpdateOrders,
+      (isUpdateOrders) => {
+        if (isUpdateOrders && !this.state.refreshing) {
+          setTimeout(() => this._getData(0, true));
+        }
+      },
+    );
   }
 
   render() {