+++ src/components/payment/Confirm/components/index.js
@@ -1,4 +1,5 @@
 import PaymentMethodSection from './PaymentMethodSection';
 import DeliverySection from './DeliverySection';
+import StoreInfo from './StoreInfo';
 
-export {PaymentMethodSection, DeliverySection};
+export {PaymentMethodSection, DeliverySection, StoreInfo};